requirements.txt
fastapi
uvicorn
python-multipart
pandas
spacy
pdfplumber
python-docx
openpyxl
